+++ model/miner.go
@@ -9,7 +9,7 @@ type Miner struct {
 	SectorSize      *uint64   `json:"sector_size"`
 	RawBytePower    *uint64   `json:"raw_byte_power"`
 	QualityAdjPower *uint64   `json:"quality_adj_power"`
-	RelativePower   *float64  `json:"relative_power"`
+	RelativePower   *float32  `json:"relative_power"`
 	Score           *uint32   `json:"score"`
 	CreatedAt       time.Time `json:"-"`
 	UpdatedAt       time.Time `json:"-"`
